Adapting function to use database from Firebase DB
It looks like a big change, but it is only change the localstorage to
firebase DB. Now the function gets array of inputs state from firebase
(const arrayOfInputsState = data.val();), next the modified table is sent back to the database(lines: 48-59).
Sending back array works like this: first the old table is removed from the
DB, then set a new array. The significant difference is that the function logic handling the
input state change is in the promise(arrayOfHabitsRef.once). This is necessary, only then is data access from firebase available